requirements.txt
docker
requests
flake8
beautifulsoup4
crewai
crewai[tools]
python-dotenv
